// src/shared/types/constants.ts
export enum DefaultPalette {
    SUCCESS = "#2AA952",
    ERROR = "#F01F0E",
    WHITE = "#FFFFFF",
    BACKGROUND = "#F9F9F9",
    PRIMARY = "#DB3022",
    GRAY = "#9B9B9B",
    BLACK = "#222222"
}

export enum Metropolis {
    BOLD = "MetropolisBold",
    MEDIUM = "MetropolisMedium",
    REGULAR = "MetropolisRegular"
} 